Print root-package relative path of the file where a failure happens
Issue #20 is about printing the absolute path and not the
package-relative path as this commit implements. We decided to use the
root-package relative path as the absolute path exposes details about
where you store your files and also makes the spec for `Spec` itself
difficult.

The root-package is the one where you are running `savi build` or `savi
run` from.

Related issues: #20